renaming component
